requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
Authlib==0.15.5
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.1
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
distlib==0.3.1
Django==3.1.5
dnspython==2.1.0
email-validator==1.1.3
filelock==3.0.12
Flask==2.0.2
Flask-BabelEx==0.9.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Secrets==0.1.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
fonttools==4.28.5
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
isort==5.7.0
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
jsonify==0.5
kiwisolver==1.3.2
lazy-object-proxy==1.4.3
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
mysql-connector==2.2.9
mysql-connector-python==8.0.27
numpy==1.21.3
packaging==21.3
pandas==1.3.5
passlib==1.7.4
Pillow==9.0.0
protobuf==3.19.1
pycparser==2.21
pygame==2.1.2
PyJWT==2.3.0
pylint==2.6.0
pymongo==4.0.1
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2020.5
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.15.0
sklearn==0.0
soupsieve==2.3.1
speaklater==1.3
SQLAlchemy==1.4.25
sqlparse==0.4.1
threadpoolctl==3.1.0
toml==0.10.2
urllib3==1.26.8
virtualenv==20.4.0
virtualenvwrapper-win==1.2.6
Werkzeug==2.0.2
wrapt==1.12.1
WTForms==2.3.3
